refactor: extract gas fee calc and USDT reset
